requirements.txt
certifi==2019.3.9
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
pyzbar==0.1.8
six==1.12.0
